replace advance_chebyshevs by next_chebyshev; rescale Cheb init
Two big things and a few small in this commit. (I know that's bad
form.)

1. advance_chebyshevs was (I think!) buggy. I've replaced it by a much
simpler function that just returns the next chebyshev function,
together with code in all_single_mu (but not all_double_mu!) to use
that function and do the replacement.

2. Our MPSs were awfully big (in the since of Frobenius norm), and it
was making me nervous---and moreover in the conductivity case we need
to divide by something that turns out to be exactly the 2^L initial
order-of-magnitude normalization anyway. The Chebyshev recursion
relation is linear in its initial state, so I just normalize that.

Note that I haven't propagated these changes to double-mu or to
conductivity.cc (though it's not clear if I'll need to propagate to
the latter).