basic verification
"make verification" now runs dos for a variety of lengths, then runs a
julia script that plots the bond dimension as a function of chebyshev
order and the singular values of a selection of Chebyshev
polynomials. It puts the data files and plots in a directory
    verification-(DATE)-(COMMIT) .
Takes a while to run.